refactor: 패키지 구조 재구성
