+++ src/main/java/com/goorm/codeAdventure/domain/game/entity/Category.java
@@ -1,0 +1,28 @@
+package com.goorm.codeAdventure.domain.game.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.Getter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@Table(name = "categories")
+public class Category {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+    @OneToMany(mappedBy = "category")
+    private List<Stage> stages = new ArrayList<>();
+
+    // getters and setters
+}
