+++ src/main/java/com/goorm/codeAdventure/domain/item/dto/ItemDetailResponse.java
@@ -1,0 +1,21 @@
+package com.goorm.codeAdventure.domain.item.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ItemDetailResponse {
+    private final Long id;
+    private final String name;
+    private final Integer stockQuantity;
+    private final Integer price;
+    //
+
+
+
+
+}
