--- src/main/java/com/goorm/codeAdventure/item/dto/ItemListResponse.java
@@ -1,19 +1,0 @@
-package com.goorm.codeAdventure.item.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ItemListResponse {
-    private final Long id;
-    private final String name;
-    private final Integer stockQuantity;
-    private final Integer price;
-
-
-
-}
