--- src/main/java/com/goorm/codeAdventure/item/dto/ItemPurchaseResponse.java
@@ -1,14 +1,0 @@
-package com.goorm.codeAdventure.item.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class ItemPurchaseResponse {
-    private final Integer usedCoin;
-    private final String  message;
-
-}
