--- src/main/java/com/goorm/codeAdventure/item/repository/ItemRepository.java
@@ -1,10 +1,0 @@
-package com.goorm.codeAdventure.item.repository;
-
-import com.goorm.codeAdventure.item.domain.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ItemRepository extends JpaRepository<Item,Long> {
-
-
-}
