--- src/main/java/com/goorm/codeAdventure/problem/domain/AttemptResult.java
@@ -1,16 +1,0 @@
-package com.goorm.codeAdventure.problem.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum AttemptResult {
-    SUCCESS("성공"),
-    FAIL("실패"),
-    NOT_ATTEMPTED("미시도");
-
-    private final String message;
-
-    AttemptResult(String message) {
-        this.message = message;
-    }
-}
