--- src/main/java/com/goorm/codeAdventure/problem/domain/SupportedLanguage.java
@@ -1,27 +1,0 @@
-package com.goorm.codeAdventure.problem.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "supported_languages")
-public class SupportedLanguage {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "programming_language_id")
-    private ProgrammingLanguage programmingLanguage;
-
-    @ManyToOne
-    @JoinColumn(name = "problem_id")
-    private Problem problem;
-}
