+++ src/test/java/com/goorm/codeAdventure/user/userService/UserServiceTest.java
@@ -1,9 +1,9 @@
 package com.goorm.codeAdventure.user.userService;
 
-import com.goorm.codeAdventure.user.domain.User;
-import com.goorm.codeAdventure.user.repository.UserRepository;
+import com.goorm.codeAdventure.domain.user.entity.User;
+import com.goorm.codeAdventure.domain.user.repository.UserRepository;
+import com.goorm.codeAdventure.domain.user.service.UserService;
 import jakarta.persistence.EntityManager;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
